refactor(processing): reorganize package structure for AsyncProcessingService and related classes
