+++ processing/src/main/java/pt/estga/processing/services/processing/ProcessingService.java
@@ -1,4 +1,4 @@
-package pt.estga.processing.services;
+package pt.estga.processing.services.processing;
 
 public interface ProcessingService {
     void processSubmission(Long submissionId);
